Fix https://github.com/StockSharp/StockSharp/commit/de2cb2e056501d80703d00c361aecbd02936832c

+++ Messages/SecurityId.cs
@@ -73,8 +73,8 @@ namespace StockSharp.Messages
 			{
 				_native = value;
 
-				if (value is int)
-					_nativeAsInt = (int)value;
+				if (value is long)
+					_nativeAsInt = (long)value;
 			}
 		}
